backend/warm_signal/requirements.txt
Django==5.2
psycopg[binary]
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

